+++ documentation/docs/release-notes.md
@@ -5,6 +5,7 @@ title: Release notes&#58;
 
 **v4.0.0**:
 
+- Improved the profile manager configuration
 
 **v3.5.0**: